chore: move impl #name (clap::Parser adapter block) into submodule
